+++ lib/src/models/weather_model.dart
@@ -7,7 +7,9 @@ class WeatherModel {
   final double pressure;
   final double humidity;
   final String iconWeather;
-  final String windstring;
+  final double windVelocity;
+  final CardinalPoint windDirection;
+  final String windDirectionDescription;
   final String descriptionWeather;
 
   /// Class constructor
@@ -18,13 +20,23 @@ class WeatherModel {
     this.pressure,
     this.humidity,
     this.iconWeather,
-    this.windstring,
+    this.windVelocity,
+    this.windDirection,
+    this.windDirectionDescription,
     this.descriptionWeather,
   });
 
   /// Static method that returns an instance of the class from the json provided
   static WeatherModel fromJson(dynamic json) {
     final data = json['data'];
+    var windString = data['windstring'];
+    var beginIndex = windString.indexOf('Velocidad') + 9;
+    var endIndex = windString.indexOf('m/s');
+    var windVelocityString = windString.substring(beginIndex, endIndex).trim();
+    var windVelocity = double.parse(windVelocityString) * 3.6;
+    beginIndex = endIndex + 3;
+    endIndex = windString.length;
+    var windDirectionDesc = windString.substring(beginIndex, endIndex).trim();
     return WeatherModel(
       cityName: data['cityName'],
       dt: WeatherDateModel.fromJson(data['dt']),
@@ -32,7 +44,9 @@ class WeatherModel {
       pressure: double.parse(data['pressure'].toString()),
       humidity: double.parse(data['humidity'].toString()),
       iconWeather: data['iconWeather'],
-      windstring: data['windstring'],
+      windVelocity: windVelocity,
+      windDirection: _parseDirection(windDirectionDesc),
+      windDirectionDescription: windDirectionDesc,
       descriptionWeather: data['descriptionWeather'],
     );
   }
@@ -45,10 +59,36 @@ class WeatherModel {
     result.write('Timestamp: ${dt.date}\n');
     result.write('Humidity: ${humidity}%\n');
     result.write('Pressure: ${pressure} hpa\n');
-    result.write('Wind: ${windstring}\n');
+    result.write('Wind Velocity: ${windVelocity} Km/h\n');
+    result.write('Wind Direction: ${windDirection}\n');
+    result.write('Wind Direction Description: ${windDirectionDescription}\n');
     result.write('Description: ${descriptionWeather}\n');
     result.write('Image Link: ${iconWeather}');
     return result.toString();
+  }
+
+  static CardinalPoint _parseDirection(String input) {
+    var direction = input.split(' ')[0].toLowerCase().trim();
+    switch (direction) {
+      case 'norte':
+        return CardinalPoint.North;
+      case 'noreste':
+        return CardinalPoint.Northeast;
+      case 'este':
+        return CardinalPoint.East;
+      case 'sureste':
+        return CardinalPoint.Southeast;
+      case 'sur':
+        return CardinalPoint.South;
+      case 'suroeste':
+        return CardinalPoint.Southwest;
+      case 'oeste':
+        return CardinalPoint.West;
+      case 'noroeste':
+        return CardinalPoint.Northwest;
+      default:
+        return CardinalPoint.North;
+    }
   }
 }
 
@@ -70,4 +110,15 @@ class WeatherDateModel {
       timezone: json['timezone'],
     );
   }
+}
+
+enum CardinalPoint {
+  North,
+  Northeast,
+  East,
+  Southeast,
+  South,
+  Southwest,
+  West,
+  Northwest,
 }